+++ packages/contracts-ts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.4.4](https://github.com/cowprotocol/cow-sdk/compare/sdk-contracts-ts-v0.4.3...sdk-contracts-ts-v0.4.4) (2025-11-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-config bumped to 0.3.0
+
 ## [0.4.3](https://github.com/cowprotocol/cow-sdk/compare/sdk-contracts-ts-v0.4.2...sdk-contracts-ts-v0.4.3) (2025-10-30)